refactor: create export to csv helper
